requirements.txt
flask
mysql-connector
pytest
flask_cors
mysql-connector-python
pyjwt
flask-pyjwt
